requirements.txt
python-twitter
appdirs==1.4.3
dj-database-url==0.4.2
Django==1.11
gunicorn==19.7.1
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
whitenoise==3.3.0
social-auth-app-django>=1.2.0
